renaming and splitting stuff up
