+++ tools/testdata/command-r-plus.gotmpl
