+++ tools/testdata/firefunction.gotmpl
