+++ tools/testdata/firefunction.out
