+++ tools/testdata/llama3-groq-tool-use.out
