+++ tools/testdata/llama3.2.gotmpl
