+++ tools/testdata/mistral.gotmpl
