+++ tools/testdata/nemotron.gotmpl
