+++ tools/testdata/qwen2.5-coder.gotmpl
