+++ tools/testdata/qwen3.gotmpl
