+++ tools/testdata/xlam.gotmpl
